refactor: 알림 허용 버튼으로 수정
